requirements.txt
aiogram
aiosqlite
python-dotenv
cryptography
APScheduler
jdatetime
qrcode
